% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_distribution.R
\name{read_length_distribution}
\alias{read_length_distribution}
\title{Read length distribution plots corresponding to mutationFileFormat files}
\usage{
read_length_distribution(dir, end_break = 5, plot_layout = c(3, 3),
  cols = c("red", "green", "blue"), cex_legend = 0.5, cex.main = 0.5)
}
\arguments{
\item{dir}{The directory housing all the mutationFileFormat files.}

\item{end_break}{The position on the read from the end of the fragment such that if
C->T change occurs beyond that end break, it is considered in the third class of
reads (potentially damaged reads)}

\item{plot_layout}{The layout of the plot with multiple images (deafult is a 3x3 layout)}

\item{cols}{The color profiles of the three plots. Defaults to "red", "green" and "blue".}

\item{cex_legend}{The cex of the legend.}

\item{cex_main}{The cex of the title of the figure.}
}
\value{
The read length distribution plots for each of the mutationFileFormat files
in the folder
}
\description{
For each file (of mutationFileFormat type) in a directory, we plot
the read length distribution under three settings - all reads, reads containing
C to T mutations and reads containing C to T towards the ends of the fragments.
}
\keyword{read_length}
